fix: redis config on serverToken
